saving progress commit before pull
