+++ src/pages/HomePage.tsx
@@ -1,5 +1,9 @@
 // type Props = {}
 
 export const HomePage = () => {
-  return <div className="bg-green-500 w-full h-full">HomePage</div>;
+  return (
+    <div className="bg-green-500 w-full h-full">
+      Welcome to Nice Gadgets store!
+    </div>
+  );
 };
